refactor(core): move css parser back to core crate
