+++ crates/core-support/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2018"
 license = "MIT/Apache-2.0"
 repository = "https://github.com/elvisjs/elvis/tree/master/crates/core-support"
 homepage = "https://github.com/elvisjs"
-documentation = "https://docs.rs/elvis-core-support"
+documentation = "https://docs.rs/core-support"
 description = """
 Definition of the `#[elvis]` attribute, an internal dependency
 """
@@ -17,4 +17,3 @@ proc-macro = true
 [dependencies]
 quote = "1.0.7"
 syn = "1.0.33"
-proc-macro2 = "1.0.19"
