+++ crates/web/src/style/mod.rs
@@ -3,7 +3,4 @@
 mod parser;
 mod sheet;
 
-pub use self::{
-    parser::{parse_class, parse_style},
-    sheet::StyleSheet,
-};
+pub use self::{parser::parse_class, sheet::StyleSheet};
